+++ bidsme/Modules/base.py
@@ -28,6 +28,8 @@ import shutil
 import logging
 import json
 import re
+import numpy
+import gzip
 
 from datetime import datetime, date, time
 from collections import OrderedDict
@@ -77,7 +79,8 @@ class baseModule(abstract):
                  "_acqTime",
                  "manufacturer",
                  "isBidsValid",
-                 "encoding"
+                 "encoding",
+                 "zip"
                  ]
 
     _module = "base"
@@ -142,6 +145,8 @@ class baseModule(abstract):
         self.isBidsValid = True
         self.encoding = "ascii"
 
+        self.zip = False
+
     #############################
     # Optional virtual methodes #
     #############################
@@ -166,8 +171,36 @@ class baseModule(abstract):
         ext: str
             extention of the data file
         """
-        shutil.copy2(self.currentFile(),
-                     os.path.join(directory, bidsname + ext))
+
+        out_fname = os.path.join(directory, bidsname + ext)
+        if self.zip:
+            with open(self.currentFile(), 'rb') as f_in:
+                with gzip.open(out_fname + ".gz", 'wb') as f_out:
+                    shutil.copyfileobj(f_in, f_out)
+        else:
+            shutil.copy2(self.currentFile(), out_fname)
+
+    def _post_copy_bidsified(self,
+                             directory: str,
+                             bidsname: str,
+                             ext: str) -> None:
+        """
+        Virtual function that performs all post-copy tasks, for ex.
+        copy needed auxiliary files or change some internal values
+        on copied file. It uses same parameters as _copy_bidsified
+        and is executed just after it.
+
+        Parameters:
+        -----------
+        directory: str
+            destination directory where files should be copies,
+            including modality folder. Assured to exists.
+        bidsname: str
+            bidsified name without extention
+        ext: str
+            extention of the data file
+        """
+        pass
 
     def copyRawFile(self, destination: str) -> str:
         """
@@ -257,7 +290,10 @@ class baseModule(abstract):
         if not os.access(file, os.R_OK):
             raise PermissionError("File {} not readable"
                                   .format(file))
-        return cls._isValidFile(file)
+        try:
+            return cls._isValidFile(file)
+        except Exception:
+            return False
 
     @classmethod
     def Module(cls):
@@ -981,6 +1017,7 @@ class baseModule(abstract):
                                                     ext))
 
         self._copy_bidsified(outdir, bidsname, ext)
+        self._post_copy_bidsified(outdir, bidsname, ext)
 
         with open(os.path.join(outdir, bidsname + ".json"), "w") as f:
             js_dict = self.exportMeta()
@@ -992,7 +1029,12 @@ class baseModule(abstract):
         self.rec_BIDSvalues["filename"] = os.path.join(self.Modality(),
                                                        bidsname
                                                        + ext)
-        self.rec_BIDSvalues["acq_time"] = self.acqTime()
+        if self.acqTime() is None:
+            self.rec_BIDSvalues["acq_time"] = None
+        else:
+            self.rec_BIDSvalues["acq_time"] = self.acqTime().replace(
+                    microsecond=0,
+                    tzinfo=None)
 
         scans = os.path.join(bidsfolder,
                              self.getBidsPrefix('/'),
@@ -1493,4 +1535,6 @@ class ExtendEncoder(json.JSONEncoder):
                 return obj.decode(self.encoding)
             except UnicodeDecodeError:
                 return "<bytes>"
+        if isinstance(obj, numpy.ndarray):
+            return obj.tolist()
         return json.JSONEncoder.default(self, obj)